Finished

+++ SimchaFund.Web/Views/ContributorVM.cs
@@ -5,6 +5,6 @@ namespace SimchaFund.Web.Views
     public class ContributorVM
     {
         public Contributor Contributor { get; set; }
-        public decimal InitialDeposit { get; set; }
+        public double InitialDeposit { get; set; }
     }
 }
