+++ SimchaFund.Web/Views/GetContributorsVM.cs
@@ -3,6 +3,6 @@
     public class GetContributorsVM
     {
         public List<GetGuyAndBalance> Contributors { get; set; }
-        public decimal Total { get; set; }
+        public double Total { get; set; }
     }
 }
