+++ SimchaFund.Web/Views/GetSimchos.cs
@@ -4,7 +4,7 @@ namespace SimchaFund.Web.Views
 {
     public class GetSimchos : Simcha
     {
-        public decimal TotalCollected { get; set; }
+        public double TotalCollected { get; set; }
         public int Contributors { get; set; }
     }
 }
